base/requirements.txt
asgiref==3.7.2
Django==5.0.1
django-cors-headers==4.3.1
djangorestframework==3.14.0
gunicorn==21.2.0
Markdown==3.5.2
packaging==23.2
pytz==2024.1
sqlparse==0.4.4
whitenoise==6.6.0
